pyproject.toml
[project]
name = "g-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-pinecone>=0.2.13",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
